refactor(onboarding): extract SIGNAL_OPTIONS to break circular import
use-onboarding-actions.ts imported SIGNAL_OPTIONS from signals-step.tsx,
while signals-step.tsx imported useOnboardingActions back — a circular
dependency that only worked by accident because the imports are consumed
inside functions, not at module init. Move the pure data constant to its
own module so both files import from a leaf.
